fix : 아이템 리스트 공통 컴포넌트 구조 변경 #106
